requirements.txt
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.7
joblib==1.4.2
jsonpickle==3.2.2
kiwisolver==1.4.5
kneed==0.8.5
matplotlib==3.9.2
more-itertools==10.4.0
music21==9.1.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tk==0.1.0
urllib3==2.2.2
webcolors==24.8.0
colormath==3.0.0
scikit-image==0.24.0